refactor: 활성화 카테고리, 글(`active~`) 설정 로직 훅으로 분리
